requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
ecdsa==0.19.1
Flask==3.1.2
flask-cors==6.0.1
flask-marshmallow==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
packaging==25.0
psycopg2==2.9.11
pyasn1==0.6.1
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.3
